+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.14.5] - 2023-06-28
+### Changed
+- [Public Explorer] Better handling of syncing/expired block in integrity checks
+
+## [3.14.4] - 2023-06-28
+### Added
+- [Public Explorer] Fix reverting pending block when including transaction
+
+## [3.14.3] - 2023-06-28
+### Added
+- [Public Explorer] Integrity check is now handling expired pending blocks
+
+### Changed
+- [Public Explorer] Block sync should revert pending blocks better
+
+## [3.14.2] - 2023-06-27
+### Fixed
+- If code retrieval during tracing fails, the whole tracing fails. Fixes this.
+
 ## [3.14.1] - 2023-06-26
 ### Changed
 - Improve case when balance change request fails